requirements.txt
python-dotenv~=1.0.1
fastapi~=0.110.2
uvicorn~=0.29.0